Make the BLE code cross-platform
